nedd refactoring
