+++ project/.gitignore
@@ -28,4 +28,4 @@ target/
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 
-/src/main/resources/credential.properties
+/model/src/main/resources/credential.properties
