+++ project/appweb/pom.xml
@@ -20,7 +20,7 @@
     <dependencies>
         <dependency>
             <groupId>vasilizas</groupId>
-            <artifactId>service</artifactId>
+            <artifactId>vasilizas.service</artifactId>
             <version>1.0.0</version>
             <scope>compile</scope>
         </dependency>
