--- project/appweb/src/main/java/web/controller/AddStudent.java
@@ -1,22 +1,0 @@
-package web.controller;
-
-
-
-
-
-public class AddStudent {
-
-//    @GetMapping("/students/add")
-//    public String doGet(Model model){
-//        model.addAttribute("message", "Hello from spring MVC");
-//        return "studentForm";
-//    }
-//
-//    @PostMapping("/students/add")
-//    protected String doPost(@RequestParam String name,
-//                            @RequestParam String lastName,
-//                            @RequestParam int age) {
-//        StudentService.saveStudent(name, lastName, age);
-//        return "redirect: /school/students";
-//    }
-}
