+++ project/appweb/src/main/java/web/vasilizas/controller/AddStudent.java
@@ -1,0 +1,22 @@
+package web.vasilizas.controller;
+
+
+
+
+
+public class AddStudent {
+
+//    @GetMapping("/students/add")
+//    public String doGet(Model model){
+//        model.addAttribute("message", "Hello from spring MVC");
+//        return "studentForm";
+//    }
+//
+//    @PostMapping("/students/add")
+//    protected String doPost(@RequestParam String name,
+//                            @RequestParam String lastName,
+//                            @RequestParam int age) {
+//        StudentService.saveStudent(name, lastName, age);
+//        return "redirect: /school/students";
+//    }
+}
