+++ project/appweb/src/main/java/web/vasilizas/controller/filter/AbstractFilter.java
@@ -1,0 +1,18 @@
+package web.vasilizas.controller.filter;
+
+import javax.servlet.Filter;
+import javax.servlet.FilterConfig;
+import javax.servlet.ServletException;
+
+public abstract class AbstractFilter implements Filter {
+
+    @Override
+    public void init(FilterConfig filterConfig) throws ServletException {
+        Filter.super.init(filterConfig);
+    }
+
+    @Override
+    public void destroy() {
+        Filter.super.destroy();
+    }
+}
