+++ project/appweb/src/main/java/web/vasilizas/controller/filter/EncodingFilter.java
@@ -1,0 +1,19 @@
+package web.vasilizas.controller.filter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+import javax.servlet.annotation.WebFilter;
+import java.io.IOException;
+
+@WebFilter(urlPatterns = "/*")
+public class EncodingFilter extends AbstractFilter {
+
+       @Override
+    public void doFilter(ServletRequest req, ServletResponse resp, FilterChain chain) throws IOException, ServletException {
+        req.setCharacterEncoding("UTF-8");
+        resp.setCharacterEncoding("UTF-8");
+        chain.doFilter(req, resp);
+    }
+}
