+++ project/appweb/src/main/webapp/WEB-INF/web.xml
@@ -6,7 +6,7 @@
 
   <servlet>
     <servlet-name>MyServlet</servlet-name>
-    <servlet-class>web.TeacherInfo</servlet-class>
+    <servlet-class>web.vasilizas.TeacherInfo</servlet-class>
   </servlet>
 
   <servlet-mapping>
@@ -14,6 +14,10 @@
     <url-pattern>/start</url-pattern>
   </servlet-mapping>
 
+  <session-config>
+    <session-timeout>
+      30
+    </session-timeout>
+  </session-config>
 
-
-</web-app>
+  </web-app>
