+++ project/appweb/src/main/webapp/admin/admin.jsp
@@ -1,0 +1,37 @@
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+
+<!DOCTYPE html>
+<html lang=en>
+<head>
+    <meta charset=\UTF-8\>
+    <title>admin</title>
+</head>
+<body align=center>
+<h1 center> Admin form </h1>
+<a href="/admin/admin/teacher-info">Информация о тренерах</a>
+<p center>
+    <c></c>
+</p>
+<a href="/admin/addstudent">Добавить студента</a>
+<p center>
+    <c></c>
+</p>
+<a href="/admin/delstudent">Удалить студента</a>
+<p center>
+    <c></c>
+</p>
+<a href="/admin/addteacher">Добавить тренера</a>
+<p center>
+    <c></c>
+</p>
+<a href="/admin/delstudent">Удалить тренера</a>
+<p center>
+    <c></c>
+</p>
+<a href="/admin/allstudent">Информация о студентах</a>
+<p center>
+    <c></c>
+</p>
+<button onclick="history.back()">Go Back</button>
+</body>
+</html>
