+++ project/appweb/src/main/webapp/index.jsp
@@ -6,11 +6,19 @@
     <meta charset=\UTF-8\>
     <title>My Web Project</title>
 </head>
-<body>
-<h1  center> My Pet Project </h1>
-<a  href="/myweb/about">Узнать о проекте</a>
-<p center><c></c></p>
-<a  href="/myweb/teacher-info">Наши тренеры</a>
+<body align=center>
+<h1 center> My Web Project </h1>
+<a href="/myweb/about">Узнать о проекте</a>
+<p center>
+    <c></c>
+</p>
 
+<p center>
+    <c></c>
+</p>
+<a href="auth.jsp">Авторизация</a>
+<p center>
+<c></c>
+</p>
 </body>
 </html>
