--- project/model/src/main/java/bean/DefaultPerson.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class DefaultPerson extends Person{
-}
