--- project/model/src/main/java/repository/TeacherRepository.java
@@ -1,14 +1,0 @@
-package repository;
-
-import bean.Teacher;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public final class TeacherRepository {
-
-    public static List<Teacher> teacherList = new ArrayList<>();
-
-    private TeacherRepository() {
-    }
-}
