+++ project/model/src/main/java/vasilizas/bean/Admin.java
@@ -1,0 +1,10 @@
+package vasilizas.bean;
+
+public class Admin extends Person{
+    public Admin() {
+    }
+
+    public Admin(String name, int age) {
+        super(name, age);
+    }
+}
