+++ project/model/src/main/java/vasilizas/bean/DefaultPerson.java
@@ -1,0 +1,4 @@
+package vasilizas.bean;
+
+public class DefaultPerson extends Person{
+}
