+++ project/model/src/main/java/vasilizas/repository/StudentRepository.java
@@ -1,0 +1,14 @@
+package vasilizas.repository;
+
+import vasilizas.bean.Student;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public final class StudentRepository {
+
+    public static List<Student> studentList = new ArrayList<>();
+
+    private StudentRepository() {
+    }
+}
