+++ project/model/src/main/java/vasilizas/repository/TeacherRepository.java
@@ -1,0 +1,14 @@
+package vasilizas.repository;
+
+import vasilizas.bean.Teacher;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public final class TeacherRepository {
+
+    public static List<Teacher> teacherList = new ArrayList<>();
+
+    private TeacherRepository() {
+    }
+}
