+++ project/service/pom.xml
@@ -11,7 +11,7 @@
     </parent>
 
     <groupId>vasilizas</groupId>
-    <artifactId>service</artifactId>
+    <artifactId>vasilizas.service</artifactId>
     <version>1.0.0</version>
 
     <properties>
