+++ project/service/src/main/java/vasilizas/myservice/security/PersonSecurity.java
@@ -1,0 +1,17 @@
+package vasilizas.myservice.security;
+
+import static vasilizas.myservice.person.MyService.log;
+
+public class PersonSecurity {
+
+    private PersonSecurity() {
+    }
+
+    public static void addLoginAndPassword(String personType, String personName, String login, String password) {
+        switch (personType) {
+            case "Student" -> StudentSecurity.security(personName, login, password);
+            case "Teacher" -> TeacherSecurity.security(personName, login, password);
+            default -> log.warn("Incorrect variable entered");
+        }
+    }
+}
