+++ project/service/src/main/java/vasilizas/myservice/security/Security.java
@@ -1,0 +1,8 @@
+package vasilizas.myservice.security;
+
+public interface  Security {
+
+      static void security(String personName, String login, String password){}
+
+      static void getPassword(String personName, String login){}
+}
